fix: imported all dynamic floating styles to stable
