+++ packages/react/src/components/Tooltip/Tooltip.featureflag.stories.js
@@ -1,0 +1,75 @@
+/**
+ * Copyright IBM Corp. 2016, 2023
+ *
+ * This source code is licensed under the Apache-2.0 license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import React from 'react';
+import { Information } from '@carbon/icons-react';
+import { Tooltip } from '../Tooltip';
+import { WithFeatureFlags } from '../../../.storybook/templates/WithFeatureFlags';
+
+import './story.scss';
+
+// eslint-disable-next-line storybook/csf-component
+export default {
+  title: 'Components/Tooltip/Feature Flag',
+  component: Tooltip,
+  tags: ['!autodocs'],
+  decorators: [
+    (Story) => (
+      <WithFeatureFlags>
+        <Story />
+      </WithFeatureFlags>
+    ),
+  ],
+};
+
+export const FloatingStyles = (args) => {
+  const tooltipLabel =
+    'Scroll the container up, down, left or right to observe how the tooltip will automatically change its position in attempt to stay within the viewport. This works on initial render in addition to on scroll.';
+  return (
+    <div
+      style={{
+        width: '100%',
+        display: 'flex',
+        justifyContent: 'center',
+      }}>
+      <Tooltip label={tooltipLabel} align={args.align}>
+        <button className="sb-tooltip-trigger" type="button">
+          <Information />
+        </button>
+      </Tooltip>
+    </div>
+  );
+};
+
+FloatingStyles.args = {
+  align: 'bottom',
+};
+
+FloatingStyles.argTypes = {
+  align: {
+    options: [
+      'top',
+      'top-start',
+      'top-end',
+
+      'bottom',
+      'bottom-start',
+      'bottom-end',
+
+      'left',
+      'left-end',
+      'left-start',
+
+      'right',
+      'right-end',
+      'right-start',
+    ],
+    control: {
+      type: 'select',
+    },
+  },
+};
